refactor: 패키지 구조 변경 - support 패키지 생성
