+++ src/main/java/io/hhplus/reserve/concert/domain/ConcertSeat.java
@@ -1,8 +1,8 @@
 package io.hhplus.reserve.concert.domain;
 
-import io.hhplus.reserve.common.domain.BaseEntity;
-import io.hhplus.reserve.common.exception.BusinessException;
-import io.hhplus.reserve.common.exception.ErrorCode;
+import io.hhplus.reserve.support.domain.BaseEntity;
+import io.hhplus.reserve.support.domain.exception.BusinessException;
+import io.hhplus.reserve.support.domain.exception.ErrorCode;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
