+++ src/main/java/io/hhplus/reserve/point/domain/PointHistory.java
@@ -1,6 +1,6 @@
 package io.hhplus.reserve.point.domain;
 
-import io.hhplus.reserve.common.domain.BaseEntity;
+import io.hhplus.reserve.support.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
