+++ src/main/java/io/hhplus/reserve/support/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package io.hhplus.reserve.common.domain;
+package io.hhplus.reserve.support.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
