+++ src/main/java/io/hhplus/reserve/support/domain/exception/ApiControllerAdvice.java
@@ -1,4 +1,4 @@
-package io.hhplus.reserve.common.exception;
+package io.hhplus.reserve.support.domain.exception;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.extern.slf4j.Slf4j;
